track cache logic example
only in the bogie, should be cloned to AbstractTrains.
in theory we can enlarge the hitbox of the TileTCRail to cover the entire track piece, the entity can then get all the tile entities at the current block, and if that list contains the last used, then continue using the last used, otherwise set last used to the first new one found.

This would obsolete TileTCRailGag completely and dramatically reduce load on pathfinding logic from less world calls.
switch tracks would likely need to be two tiles, one for the active path, and others for the inactive paths, likely dynamically deleting and creating tiles as states change...?