more work on pathfinding
i realized that sometimes you dont enter a track from a host block, and the logic didn't like that. especially in transition from vanilla track, so, now there's compensation for that scenario.
also redefined movement to function the same on host and gag blocks, needs testing but in theory should be more reliable. should.